Convert more adapter files to TypeScript
