+++ packages/adapter-netlify/.gitignore
@@ -1,3 +1,4 @@
 .DS_Store
 node_modules
 render.js
+render.d.ts
