+++ packages/adapter-netlify/rollup.config.js
@@ -3,7 +3,7 @@ import commonjs from '@rollup/plugin-commonjs';
 import typescript from 'rollup-plugin-typescript2';
 
 export default {
-	input: 'src/index.ts',
+	input: 'src/render.ts',
 	output: {
 		file: 'render.js',
 		format: 'cjs'
