+++ packages/adapter-node/rollup.config.js
@@ -2,8 +2,20 @@ import { nodeResolve } from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
 import typescript from 'rollup-plugin-typescript2';
 
-export default {
+export default [{
 	input: 'src/index.ts',
+	output: {
+		file: 'index.js',
+		format: 'cjs'
+	},
+	plugins: [
+		nodeResolve(),
+		commonjs(),
+		typescript()
+	],
+	external: require('module').builtinModules
+}, {
+	input: 'src/server.ts',
 	output: {
 		file: 'server.js',
 		format: 'cjs'
@@ -14,4 +26,4 @@ export default {
 		typescript()
 	],
 	external: require('module').builtinModules
-};
+}];
