+++ packages/adapter-static/.gitignore
@@ -1,3 +1,6 @@
 .DS_Store
 node_modules
+index.js
+index.d.ts
 server.js
+server.d.ts
