+++ packages/adapter-vercel/.gitignore
@@ -1,2 +1,6 @@
 .DS_Store
 node_modules
+index.js
+index.d.ts
+server.js
+server.d.ts
